Feature checkout

+++ BE/src/main/java/com/example/E_Learning/DTO/Response/CourseDetailResponse.java
@@ -18,7 +18,7 @@ public class CourseDetailResponse {
 	private String description;
 	private String thumbnailUrl;
 	private Long price;
-	private Integer discountPercent;
+	private Long discountPercent;
 	private Integer studentCount;
 	private LocalDateTime createdAt;
 	private CategoryResponse category;
