+++ BE/src/main/java/com/example/E_Learning/DTO/Response/CourseResponse.java
@@ -19,7 +19,7 @@ public class CourseResponse {
 	private String title;
 	private String description;
 	private String thumbnailUrl;
-	private Integer discountPercent;
+	private Long discountPercent;
 	private BigDecimal price;
 	private LocalDateTime createdAt;
 	private Integer studentCount;
